Create custom hook for todolist component
